Updated graph to include hyperlinks to visited pages
The nodes on the graph will now link back to the respective webpages they represent for easy access.

The demo in `HierarchicalClustering.py` is now deprecated as the sample data is only text files with the filenames = webpage titles.

We can quickly fix this for demonstration purposes by hard coding the urls into the `Document` objects. I included example code for this in plot.py with the `associate_urls` function.