Building out chrome extension functionality with on-demand AJAX requests
This commit reflects an extensive overhaul of the app's functionality. The Hierarchical Clustering methodology has been modified to account for misalignments in the dimensions of the TF-IDF vector and the number of labels our dendrograms expect. Further research is required here to ensure that no duplicate URLs (or webpage titles) are being sent, otherwise the server will throw an error.

I've also overhauled documents. Since we're no longer doing POC, the document objects are now represented directly with urls, webpage titles, and text. The functionality for loading in the sample data is currently deprecated as a result.

I've also redone the frontend by adding JS to perform AJAX fetch from the local server. Further development on the frontend is required (I despise CSS).

More changes were made to the WebScraper class to enable quicker parsing of webpage data. The basic idea is that the client will send the list of URLs to the server, which will then perform all the calculations and plotting and send back the dendrogram.

Will work on these parts more tomorrow.