Refactored the frontend display to be more dynamic
Refactored the logic for displaying the graph so that now before the user queries the server they will see centered text indicating that they need to query with the API. Once this is done, the text is removed and the display setting of the container are reset to facilitate proper display of the graph returned.

There's still a minor bug where a white line appears to the right of the graph that I'll work on resolving later. Overall the frontend feels much more dynamic with these changes made.

I also added different sized icons to make them display correctly in Chrome.