Updated API calls to be asynchronous on frontend and refactored the caching of the plots in the backend
There was an issue where the plot, despite being created without issues on the backend, would fail to display in the frontend for the user to see. I figured that this was an issue with the sync of the API calls back and forth between the frontend and backend, so I started refactoring the way the API calls were made to function in a more synchronous way with Promises and awaits.

I also changed the methodology for loading in the graphs on the frontend. Before, the figure would get cached and the frontend would check availability and *update* when the graph was available. But that resulted in having to reopen the extension in order to see. So now, in the backend, once a graph has been successfully created the layout of the dash app is updated with the new figure.

After refactoring these things the extension has become significantly faster and can now update on command.

Still need to figure out how to do better preprocessing and handle exceptions for when the dendrogram cannot be created, and need to set up the clickable URLs on the nodes.