Added demo, restructured directory, fixed bug where the iframe.src would not update
In this commit I added a demo.py file that will open sample webpages for the user and start the server so they can use Historian plug-and-play. I also began restructuring the directory for our Python code and will do the same for the JS code in the next commit. Finally, I fixed a bug where the iframe on the frontend would not update after the user submitted their search history by hard coding the server address into the iframe only after the dendrogram has been created instead of latently having the server address visible to the iframe upon initialization.